Filling historical RIPE ressources attributions

+++ ComputationVM/Step2_Fill_RIRs_database/2_fill_historical_AFRINIC_v3_IPv6prefixesonly.py
@@ -27,58 +27,71 @@ def get_immediate_subdirectories(dir):
     return [name for name in os.listdir(dir)
             if os.path.isdir(os.path.join(dir, name))]
 
-
 ## Launch a script located in the same folder which removes duplicates
 #os.system("python clean_RIRs_DBs.py")
 
-
-## RIR data extraction has been performed several times in the literature: see for instance the C++ code
-## https://code.google.com/p/ip-countryside/source/browse/trunk/getDBs.sh?r=4.
 
 ### sleep a random time before starting any operation
 value = random.randint(0,10)
 time.sleep(value * random.randint(0,3))
 
+
+## RIR data extraction has been performed several times in the literature: see for instance the C++ code
+## https://code.google.com/p/ip-countryside/source/browse/trunk/getDBs.sh?r=4.
 website = "ftp://ftp.afrinic.net/pub/stats/afrinic/"
 folder_download = "ftp.afrinic.net/pub/stats/afrinic/"
 
 print 'Download all the folders of allocation'
     
-command = """ wget -H -r --level=2 -k -p """ + website
+command = """ wget -N -H -r --level=2 -k -p """ + website
 print '\n\n command =', command
-#os.system(command)
-
-
-##decompressing the folders
-command = """gunzip -r """ + folder_download + "*/*.gz"
-os.system(command)
-
-command = """bzip2 -dk """ + folder_download + "*/*.bz2"
-os.system(command)
-
-## removing unuseful files (mostly compressed ones)
-command = """ rm -f """ + folder_download + "*/*.bz2"
-os.system(command)
-
-command = """ rm -f """ + folder_download + "*/*.md5"
-os.system(command)
-
-command = """ rm -f """ + folder_download + "*/*.asc"
-os.system(command)
-
-command = """ rm -f """ + folder_download + "*/*.gz.bck"
-os.system(command)
-
-command = """ rm -f """ + folder_download + "*/*.asc.gz"
-os.system(command)
-
-command = """ rm -f """ + folder_download + "*/*.md5.gz"
 os.system(command)
 
 
+## decompress
+if glob.glob(folder_download + "*/*.gz"):
+    command = """gunzip -r """ + folder_download + "*/*.gz"
+    os.system(command)
 
-## Which are the folders after download:
-folders = get_immediate_subdirectories(folder_download)
+
+if glob.glob(folder_download + "*/*.bz2"):
+    command = """bzip2 -dk """ + folder_download + "*/*.bz2"
+    os.system(command)
+    
+    ## remove
+    command = """ rm -f """ + folder_download + "*/*.bz2"
+    os.system(command)
+
+
+## Remove unuseful files.
+if glob.glob(folder_download + "*/*.md5"):
+    command = """ rm -f """ + folder_download + "*/*.md5"
+    os.system(command)
+
+if glob.glob(folder_download + "*/*.asc"):
+    command = """ rm -f """ + folder_download + "*/*.asc"
+    os.system(command)
+
+if glob.glob(folder_download + "*/*.gz.bck"):
+    command = """ rm -f """ + folder_download + "*/*.gz.bck"
+    os.system(command)
+
+
+if glob.glob(folder_download + "*/*.asc.gz"):
+    command = """ rm -f """ + folder_download + "*/*.asc.gz"
+    os.system(command)
+
+if glob.glob(folder_download + "*/*.md5.gz"):
+    command = """ rm -f """ + folder_download + "*/*.md5.gz"
+    os.system(command)
+
+
+## Which are the folders containing useful information after download:
+folders = []
+List_possible_folder_download = ['ftp.afrinic.net/pub/stats/afrinic/', 'ftp.afrinic.net/pub/stats/']
+for folder_download in List_possible_folder_download:
+    folders += get_immediate_subdirectories(folder_download)
+
 
 for folder in folders:
     print folder
